requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.12.3
certifi==2026.1.4
click==8.3.1
discord.py==2.6.4
fastapi==0.115.6
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
multidict==6.7.0
propcache==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.41.3
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
uvicorn==0.34.0
yarl==1.22.0
